Refactor /f top and attempt to fix bal tabulation
Refactor /f top to be more extensible to additional metrics in the
future - removes the wild chain of if statements and duplicated code
into a normalized indirect pattern. This should allow for easier
maintinance and customization on how each metric is printed in the
future.

Also attempts to fix a problem in /f top bal where an economy plugin may
not properly map a player's uuid to their bank account - first lookup
the OfflinePlayer then use that to query Vault. If this does turn out to
be the problem, this is probably a widespread issue throughout Factions
(both in querying and setting bank values).

Fixes #1295 (?)